ADD : 보스 체력바 프로토타입 추가

+++ Assets/Scripts/Enemy/Boss/BossStatsData.cs
@@ -13,6 +13,6 @@ public class BossStatsData : ScriptableObject
         NORMAL, TIME_ATTACK
     }
 
-    [SerializeField] public float hp;
+    [SerializeField] public int hp;
     [SerializeField] public HealthType healthType;
 }
